new and improved loader logic
